+++ ui/src/tempVars/components/MetaQueryTemplateBuilder.tsx
@@ -4,7 +4,7 @@ import {getDeep} from 'src/utils/wrappers'
 
 import {ErrorHandling} from 'src/shared/decorators/errors'
 import TemplateMetaQueryPreview from 'src/tempVars/components/TemplateMetaQueryPreview'
-import {hydrateTemplate} from 'src/tempVars/utils/hydrate'
+import {hydrateTemplate} from 'src/tempVars/utils/graph'
 import {isInvalidMetaQuery} from 'src/tempVars/parsing'
 
 import {TemplateBuilderProps, RemoteDataState} from 'src/types'
