+++ ui/src/tempVars/components/TemplateControlBar.tsx
@@ -7,7 +7,7 @@ import TemplateControl from 'src/tempVars/components/TemplateControl'
 import OverlayTechnology from 'src/reusable_ui/components/overlays/OverlayTechnology'
 import TemplateVariableEditor from 'src/tempVars/components/TemplateVariableEditor'
 import Authorized, {EDITOR_ROLE} from 'src/auth/Authorized'
-import {graphFromTemplates} from 'src/tempVars/utils/hydrate'
+import {graphFromTemplates} from 'src/tempVars/utils/graph'
 
 import {Template, TemplateValue, Source} from 'src/types'
